Hammer Wrapper and updated sync_capsule_tools_repos_to_upgrade task
